Support Repla JSON for plugins (#74)
* Clean up warnings

* Add placeholder files

* Start setting up XMLPlugin

* Setup JSON plugin

* Try JSON plugin first

* Bump SodaFountain

* Setup JSON plugin test

* Rename to PluginKind

* Rename to plugin kind

* Start implementing plugin type

* Rename plugin to base plugin

* Revert "Rename plugin to base plugin"

This reverts commit fac877f84f6bb33153f8ff741079324bc89061c7.

* Work on compile errors

* Fix more compile errors

* Work on setting up BasePlugin

* Fix some compile errors

* Fix some compile errors

* Fix a compile error

* Start switching back to abstract super class

* Fix some compile errors

* Fix compile error

* Fix some compile errors

* Work on compile errors

* Work on some compile errors

* Run autocorrect

* Fix some compile errors

* Fix some compile errors

* Align parameters

* Pass in bundle URL and path

* Swap directory path to path and use resource path

* Fix some compile errors

* Fix some compile errors

* Move comment

* Fix initializer parameter names

* Fix some compile errors

* Fix some compile errors

* Fix some compile errors

* Fix some compile errors

* Fix some compile errors

* Fix some compile errors

* Fix some compile errors

* Add commandPath to Plugin base class

* Clean up

* Migrate one property

* Re-organize properties

* Remove superfluous variable initialization

* Remove assert for changing editable

* Fix bundle paths

* Enable test

* Setup JSON plugin info

* Add thread check

* Setup failing test

* Finish serialization test

* Work on init

* Add missing key

* Setup default values

* Run autocorrect

* Disable rule

* Run autocorrect

* Work on loading JSON plugin

* Finish writing JSONPlugin

* Improve log messages

* Make load Node JSON plugin pass

* Run autocorrect

* Add fallback resource path for plugins

* Bump test data

* Setup file extension property test

* Use XCTFail

* Add comment

* Lay groundwork for forcing XML plugins

* Use force XML flag

* Force all XML plugins

* Fallback to JSON after XML with force XML

* Make first test that runs for both XML and JSON

* Fix compile error

* Add missing case

* Simplify use XML

* Setup file extension test with make plugin manager

* Tweak default plugin info

* Finish adding file extensions

* Test actual file extensions

* Add comment for comparing file extensions

* Simplify boolean test

* Remove superfluous constant

* Check for multiple info path components

* Add first repla.json test

* Work on adding and removing JSON info

* Clean up

* Add path checks

* Fix info dictionary path

* Check for correct callback

* Setup JSON write

* Work on setting up writing plugin info

* Use pretty print

* Implement JSON plugin save

* Setup temporary two plugins test case

* Continue setting up two plugins test case

* Finish setting up two plugins test case

* Setup test to run on both plugin types

* Fix runtime directory creation error

* Add type assert check

* Update project settings

* Refactor info contents to support JSON plugins

* Add loop for states

* Setup failing edit plugin test

* Add plugin type check

* Add comment

* Fix typo

* Make new temporary plugins manager tests

* Set test plugin editable

* Clean up created user plugins directory

* Add URL to description

* Fix some test errors

* Fix test failure

* Fix test failure

* Fix typo

* Bump dependencies

* Fix test failure and add comment

* Fix test failure

* Fix test failure

* Fix fallback XML loading

* Fix some test failures

* Run autocorrect

* Fix linter warnings

* Update dependencies

+++ SodaPopTests/FilesAndPluginsDirectoryManagerTests.swift
@@ -131,6 +131,10 @@ extension FilesAndPluginsDirectoryManagerTests {
                 fileWasCreatedOrModifiedExpectation.fulfill()
             }
         })
+        guard isTemporaryItem(atPath: path) else {
+            XCTFail()
+            return
+        }
         OutOfTouch.createFile(atPath: path, handler: nil)
         waitForExpectations(timeout: defaultTimeout, handler: nil)
     }
@@ -142,6 +146,10 @@ extension FilesAndPluginsDirectoryManagerTests {
                 directoryWasCreatedOrModifiedExpectation.fulfill()
             }
         })
+        guard isTemporaryItem(atPath: path) else {
+            XCTFail()
+            return
+        }
         OutOfTouch.createDirectory(atPath: path, handler: nil)
         waitForExpectations(timeout: defaultTimeout, handler: nil)
     }
@@ -155,6 +163,10 @@ extension FilesAndPluginsDirectoryManagerTests {
                 fileWasRemovedExpectation.fulfill()
             }
         })
+        guard isTemporaryItem(atPath: path) else {
+            XCTFail()
+            return
+        }
         OutOfTouch.removeFile(atPath: path, handler: nil)
         waitForExpectations(timeout: defaultTimeout, handler: nil)
     }
@@ -166,6 +178,10 @@ extension FilesAndPluginsDirectoryManagerTests {
                 directoryWasRemovedExpectation.fulfill()
             }
         })
+        guard isTemporaryItem(atPath: path) else {
+            XCTFail()
+            return
+        }
         OutOfTouch.removeDirectory(atPath: path, handler: nil)
         waitForExpectations(timeout: defaultTimeout, handler: nil)
     }
@@ -488,6 +504,37 @@ class FilesAndPluginsDirectoryManagerTests: TemporaryDirectoryTestCase {
         // because this could be the delete after move of a valid plugin's contents directory
         createPluginInfoDictionaryWasRemovedExpectation(forPluginPath: testPluginDirectoryPath)
         removeDirectoryWithConfirmation(atPath: testPluginContentsDirectoryPath)
+
+        // Remove the directory in the plugins directory, this should cause a callback
+        // because this could be the delete after move of a valid plugin
+        createPluginInfoDictionaryWasRemovedExpectation(forPluginPath: testPluginDirectoryPath)
+        removeDirectoryWithConfirmation(atPath: testPluginDirectoryPath)
+    }
+
+    func testDirectoryForInfoJSON() {
+        // Create a directory in the plugins directory, this should not cause a callback
+        let testPluginDirectoryPath = baseDirectoryPath.appendingPathComponent(testPluginDirectoryName)
+        createDirectoryWithConfirmation(atPath: testPluginDirectoryPath)
+
+        // Create a directory for the info JSON, this should not cause a callback
+        let testPluginInfoPath =
+            testPluginDirectoryPath.appendingPathComponent(infoPathComponent)
+        createDirectoryWithConfirmation(atPath: testPluginInfoPath)
+
+        // Clean Up Directory
+        // Remove info JSON directory, this should cause a callback
+        createPluginInfoDictionaryWasRemovedExpectation(forPluginPath: testPluginDirectoryPath)
+        removeDirectoryWithConfirmation(atPath: testPluginInfoPath)
+
+        // Create the file at the correct path
+        // Create a file for the info JSON, this should cause a callback
+        createPluginInfoDictionaryWasCreatedOrModifiedExpectation(forPluginPath: testPluginDirectoryPath)
+        createFileWithConfirmation(atPath: testPluginInfoPath)
+
+        // Clean up
+        // Remove info JSON info dictionary, this should cause a callback
+        createPluginInfoDictionaryWasRemovedExpectation(forPluginPath: testPluginDirectoryPath)
+        removeFileWithConfirmation(atPath: testPluginInfoPath)
 
         // Remove the directory in the plugins directory, this should cause a callback
         // because this could be the delete after move of a valid plugin
